IBM Data Science Capstone Project
I came up with a business problem, explained data required and the sources of data collection and the methodology used. This led me to present the results and discussions regarding the results. In this case, my business problem was to decide on a region in Singapore to start up a new food business. I used MRT Stations as regions around Singapore and used the Foursquare API to analyse venues and food places in a 500m radius of each MRT Station. Through this, I developed a k means clustering model of the MRT Stations, classified by their top 10 most common venues around the MRT Stations.