Capstone Project for IBM Data Science
In this project I aim to gather the most commonly visited places from each MRT Station and find out the categories of the venues. This will be done by using Foursquare to extract venue and venue category datas for each MRT Station around a 500m radius of the MRT Station.